Add Display and Serialize/Deserialize implementations
Added `fmt::Display` trait implementations for `DeltaAdjustment`, `DeltaPositionInfo`, and `DeltaInfo` to improve readability. Also introduced serialization and deserialization support for `DeltaInfo` and `DeltaAdjustment` using Serde for easier persistence and data transfer. Includes unit tests for validation of display formatting and JSON serialization.
