+++ README.md
@@ -18,7 +18,7 @@
  [![Dependencies](https://img.shields.io/librariesio/github/joaquinbejar/OptionStratLib)](https://libraries.io/github/joaquinbejar/OptionStratLib)
  [![Documentation](https://img.shields.io/badge/docs-latest-blue.svg)](https://docs.rs/optionstratlib)
 
- # OptionStratLib v0.3.6: Financial Options Library
+ # OptionStratLib v0.3.7: Financial Options Library
 
  ## Table of Contents
  1. [Introduction](#introduction)
