+++ examples/examples_strategies_delta/src/bin/strategy_long_straddle.rs
@@ -52,10 +52,7 @@ fn main() -> Result<(), Box<dyn Error>> {
 
     info!("Delta:  {:#?}", strategy.delta_neutrality()?);
     info!("Delta Neutral:  {}", strategy.is_delta_neutral());
-    info!(
-        "Delta Suggestions:  {:#?}",
-        strategy.delta_adjustments()?
-    );
+    info!("Delta Suggestions:  {:#?}", strategy.delta_adjustments()?);
 
     Ok(())
 }
