+++ examples/examples_strategies_delta/src/bin/strategy_short_butterfly_spread.rs
@@ -56,10 +56,7 @@ fn main() -> Result<(), Box<dyn Error>> {
 
     info!("Delta:  {:#?}", strategy.delta_neutrality()?);
     info!("Delta Neutral:  {}", strategy.is_delta_neutral());
-    info!(
-        "Delta Suggestions:  {:#?}",
-        strategy.delta_adjustments()?
-    );
+    info!("Delta Suggestions:  {:#?}", strategy.delta_adjustments()?);
 
     Ok(())
 }
