+++ src/greeks/mod.rs
@@ -90,5 +90,5 @@ mod utils;
 
 pub use equations::{delta, gamma, rho, rho_d, theta, vega, Greek, Greeks};
 pub(crate) use utils::calculate_d_values;
-pub use utils::{big_n, d1, d2, n};
 pub use utils::calculate_delta_neutral_sizes;
+pub use utils::{big_n, d1, d2, n};
