+++ src/strategies/bear_call_spread.rs
@@ -2209,7 +2209,7 @@ mod tests_delta {
                 quantity,
                 strike,
                 option_style,
-                side
+                side,
             } => {
                 assert_pos_relative_eq!(*quantity, delta, Positive(DELTA_THRESHOLD));
                 assert_pos_relative_eq!(*strike, k, Positive(DELTA_THRESHOLD));
@@ -2251,7 +2251,7 @@ mod tests_delta {
                 quantity,
                 strike,
                 option_style,
-                side
+                side,
             } => {
                 assert_pos_relative_eq!(*quantity, delta, Positive(DELTA_THRESHOLD));
                 assert_pos_relative_eq!(*strike, k, Positive(DELTA_THRESHOLD));
@@ -2342,7 +2342,7 @@ mod tests_delta_size {
                 quantity,
                 strike,
                 option_style,
-                side
+                side,
             } => {
                 assert_pos_relative_eq!(*quantity, delta, Positive(DELTA_THRESHOLD));
                 assert_pos_relative_eq!(*strike, k, Positive(DELTA_THRESHOLD));
@@ -2383,7 +2383,7 @@ mod tests_delta_size {
                 quantity,
                 strike,
                 option_style,
-                side
+                side,
             } => {
                 assert_pos_relative_eq!(*quantity, delta, Positive(DELTA_THRESHOLD));
                 assert_pos_relative_eq!(*strike, k, Positive(DELTA_THRESHOLD));
