+++ src/strategies/bear_put_spread.rs
@@ -2249,7 +2249,7 @@ mod tests_delta {
                 quantity,
                 strike,
                 option_style,
-                side
+                side,
             } => {
                 assert_pos_relative_eq!(*quantity, delta, Positive(DELTA_THRESHOLD));
                 assert_pos_relative_eq!(*strike, k, Positive(DELTA_THRESHOLD));
@@ -2290,7 +2290,7 @@ mod tests_delta {
                 quantity,
                 strike,
                 option_style,
-                side
+                side,
             } => {
                 assert_pos_relative_eq!(*quantity, delta, Positive(DELTA_THRESHOLD));
                 assert_pos_relative_eq!(*strike, k, Positive(DELTA_THRESHOLD));
