+++ tests/unit/strategies/delta/strategy_iron_butterfly.rs
@@ -72,7 +72,7 @@ fn test_iron_butterfly_integration() -> Result<(), Box<dyn Error>> {
             quantity,
             strike,
             option_style,
-            side, 
+            side,
         } => {
             assert_pos_relative_eq!(
                 *quantity,
