+++ tests/unit/strategies/delta/strategy_poor_mans_covered_call.rs
@@ -2,12 +2,12 @@ use optionstratlib::greeks::Greeks;
 use optionstratlib::model::types::{ExpirationDate, OptionStyle};
 use optionstratlib::strategies::delta_neutral::DeltaNeutrality;
 use optionstratlib::strategies::poor_mans_covered_call::PoorMansCoveredCall;
+use optionstratlib::strategies::DeltaAdjustment::BuyOptions;
 use optionstratlib::strategies::DELTA_THRESHOLD;
 use optionstratlib::utils::setup_logger;
 use optionstratlib::{assert_decimal_eq, assert_pos_relative_eq, pos, Positive};
 use rust_decimal_macros::dec;
 use std::error::Error;
-use optionstratlib::strategies::DeltaAdjustment::BuyOptions;
 
 #[test]
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test::wasm_bindgen_test)]
@@ -70,7 +70,7 @@ fn test_poor_mans_covered_call_integration() -> Result<(), Box<dyn Error>> {
             quantity,
             strike,
             option_style,
-            side ,
+            side,
         } => {
             assert_pos_relative_eq!(
                 *quantity,
