+++ tests/unit/strategies/delta/strategy_short_butterfly_spread.rs
@@ -74,7 +74,7 @@ fn test_short_butterfly_spread_integration() -> Result<(), Box<dyn Error>> {
             quantity,
             strike,
             option_style,
-            side, 
+            side,
         } => {
             assert_pos_relative_eq!(
                 *quantity,
